updated user hooks
