+++ src/hooks/utils/helperFunctions.js
@@ -1,7 +1,10 @@
 export const formatError = (err) => {
   let error = "";
   if (err?.response) {
-    error = err?.response?.data?.message || "An error occurred.";
+    error =
+      err?.response?.data?.message ||
+      err?.response?.data?.error ||
+      "An error occurred.";
   } else if (err?.request) {
     error = "Network error. Please try again later.";
   } else {
