+++ src/hooks/utils/useHandleResponse.js
@@ -7,7 +7,17 @@ export const useHandleResponse = () => {
     if (res?.ok) {
       dispatch(setSuccess(res?.ok));
     } else if (res?.error) {
-      dispatch(setError(res?.error));
+      dispatch(setError(parseError(res?.error)));
+    }
+  };
+
+  const parseError = (error) => {
+    if (error.includes("Unable to login")) {
+      return "Incorrect Username or Password!";
+    } else if (error.includes("duplicate key error")) {
+      return "Username already taken!";
+    } else {
+      return error;
     }
   };
