+++ src/pages/login/LogIn.js
@@ -2,14 +2,13 @@ import styles from "./LogIn.module.css";
 
 import { useState } from "react";
 
-import { useLogin } from "../../hooks/user/useLogin";
+import { useLogin } from "../../hooks/user/userApis";
 
 import Spinner from "../../components/loaders/spinner/Spinner";
 
-import { useDispatch } from "react-redux";
-import { setError, setSuccess } from "../../features/alertSlice";
 import images from "../../constants/images";
 import Button from "../../components/buttons/Button";
+import { useHandleResponse } from "../../hooks/utils/useHandleResponse";
 
 export default function LogIn() {
   const [passwordType, setPasswordType] = useState("eyePassword");
@@ -19,13 +18,7 @@ export default function LogIn() {
 
   const { login, isPending } = useLogin();
 
-  const dispatch = useDispatch();
-
-  const parseError = (error) => {
-    return error.includes("Unable to login")
-      ? "Incorrect Username or Password!"
-      : error;
-  };
+  const { handleResponse } = useHandleResponse();
 
   const showPassword = () => {
     passwordType === "eyePassword"
@@ -36,11 +29,7 @@ export default function LogIn() {
   const handleSubmit = async (e) => {
     e.preventDefault();
     const res = await login(userName, password);
-    if (res.ok) {
-      dispatch(setSuccess(res.ok));
-    } else if (res.error) {
-      dispatch(setError(parseError(res.error)));
-    }
+    handleResponse(res);
   };
 
   return (
